+++ src/main/java/com/rzodeczko/presentation/dto/HotelResponseDto.java
@@ -1,4 +1,4 @@
 package com.rzodeczko.presentation.dto;
 
-public record HotelResponseDto(Long hotelId, int capacity) {
+public record HotelResponseDto(Long hotelId, Long capacity) {
 }
